requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
httpx==0.24.1
openai==1.0.0
python-dotenv==1.0.0
groq==0.31.0
numpy==2.3.2
opencv-python==4.12.0